+++ assets/src/edit-story/components/canvas/useCanvasSelectionCopyPaste.js
@@ -60,8 +60,8 @@ function useCanvasSelectionCopyPaste( container ) {
 			const htmlContent = selectedElements
 				.map( ( { type, ...rest } ) => {
 					// eslint-disable-next-line @wordpress/no-unused-vars-before-return
-					const { Save } = getDefinitionForType( type );
-					return renderToString( <Save { ...rest } /> );
+					const { Output } = getDefinitionForType( type );
+					return renderToString( <Output { ...rest } /> );
 				} )
 				.join( '\n' );
