+++ assets/src/edit-story/elements/image/index.js
@@ -4,7 +4,7 @@
 import { PanelTypes } from '../../panels';
 export { default as Display } from './display';
 export { default as Edit } from './edit';
-export { default as Save } from './save';
+export { default as Output } from './output';
 export { default as TextContent } from './textContent';
 
 export const defaultAttributes = {
