+++ assets/src/edit-story/elements/video/index.js
@@ -3,7 +3,7 @@
  */
 import { PanelTypes } from '../../panels';
 export { default as Display } from './display';
-export { default as Save } from './save';
+export { default as Output } from './output';
 
 export const defaultAttributes = {
 	controls: false,
